+++ bw.workflow.visaapplicantsenrollment/src/main/java/sa/gov/nic/bio/bw/workflow/visaapplicantsenrollment/controllers/ReviewAndSubmitPaneFxController.java
@@ -135,9 +135,11 @@ public class ReviewAndSubmitPaneFxController extends WizardStepFxControllerBase
 		
 		lblMobileNumber.setText("+" + mobileNumber);
 		
-		GuiUtils.attachFingerprintImages(fingerprintBase64Images, ivRightThumb, ivRightIndex, ivRightMiddle,
-		                                 ivRightRing, ivRightLittle, ivLeftThumb, ivLeftIndex, ivLeftMiddle,
-		                                 ivLeftRing, ivLeftLittle);
+		GuiUtils.attachFingerprintImages(fingerprintBase64Images, null, ivRightThumb, ivRightIndex,
+		                                 ivRightMiddle, ivRightRing, ivRightLittle, ivLeftThumb, ivLeftIndex,
+		                                 ivLeftMiddle, ivLeftRing, ivLeftLittle, null,
+		                                 null, null, null,
+		                                 null, null);
 	}
 	
 	@Override